+++ adventofcode-2025/src/test/kotlin/nl/mvdr/adventofcode/adventofcode2025/day07/Part2Test.kt
@@ -3,7 +3,7 @@ package nl.mvdr.adventofcode.adventofcode2025.day07
 import nl.mvdr.adventofcode.solver.FunctionSolverTest
 import org.junit.jupiter.params.provider.Arguments
 
-class Part2Test: FunctionSolverTest<Int>(::solvePart2) {
+class Part2Test: FunctionSolverTest<Long>(::solvePart2) {
     companion object {
         @JvmStatic
         fun testSolution(): List<Arguments> {
